feat(chat): integrate marketplace model selector in quick app editor
